Extract getPeekCalls to external function
